scripts.cat.history: format json examples History.add_lead_ceremony: Moved cat.generate_lead_ceremony to History
